Fixes #30126 - use facets fact parsing
